Firebase and Firestore FINALLY FUNCTIONAL:
GameActivity.kt:
Fixed a bug in which "All" was being displayed for the game status. Remove it from the array of options.
Fixed a bug in which the age Picker filled into the platform picker's slot.
Changed ageRatingOptions && platformOptions assignings to be list as they were changed in GameModel.kt.
Changed the JSON store update, create and delete to FirebaseStore versions.

GameAdapter.kt:
Added a update function to notify the recycleview of changes as Firestore is async whilst json is synced.

FirebaseStore.tk:
Created class from scratch, FetchAll has officially taken my soul but works now.

GameListActivity.kt:
Created an adapter with an empty dataset to attach to RecycleView. A destination for firestore data to replace the empty list.

Added a reloadGames function which takes all the games from firestore and pushes them to the adapter.

GameModel.kt:
Replaced the Arrays with Lists as it kept causing crashes. Apparently Firestore dislikes Arrays. RIP ~2hrs

MainApp.kt:
Commented out GameJSONStore and replaced it with GameFirebaseStore.

This was a painful day. Goodnight :(